actu

+++ app/Http/Controllers/EmpleadoController.php
@@ -16,8 +16,13 @@ class EmpleadoController extends Controller
     public function index()
     {
         //
-        $datos['empleados']=Empleado::paginate(5);
-        return view('empleado.index', $datos);
+       // $datos['empleados']=Empleado::paginate(20);
+       // return view('empleado.index', $datos);
+
+            //
+            $datos['empleados']=Empleado::paginate(20);
+            return view('empleado.index', $datos);
+      
     }
 
     /**
@@ -39,6 +44,23 @@ class EmpleadoController extends Controller
      */
     public function store(Request $request)
     {
+
+//
+$campos=[
+    'Nombre' => 'required|string|max:100',
+    'Apellido' => 'required|string|max:100',
+    'Correo' => 'required|email',
+    'Telefono' => 'required|string|max:100',
+    'Foto' => 'required|max:10000|mimes:jpeg,png,jpg',
+
+
+];
+$mensaje=[
+'required'=>'El :attribute es requerido',
+'Foto.required' =>'La foto es requerida'
+];
+$this->validate($request, $campos, $mensaje);
+
         //introducir los datos a la basede datos
         //$datosEmpleado = request()->all();
         $datosEmpleado = request()->except('_token');
@@ -47,8 +69,9 @@ class EmpleadoController extends Controller
             $datosEmpleado['Foto']=$request->file('Foto')->store('uploads','public');
         }
         Empleado::insert($datosEmpleado);
-        return response()->json($datosEmpleado);
-    }
+        //return response()->json($datosEmpleado);
+        return redirect('empleado')->with('mensaje', 'Empleado agregado');
+        }
 
     /**
      * Display the specified resource.
@@ -84,6 +107,25 @@ return view('empleado.edit', compact('empleado'));
      */
     public function update(Request $request,  $id)
     {
+        //
+        $campos=[
+            'Nombre' => 'required|string|max:100',
+            'Apellido' => 'required|string|max:100',
+            'Correo' => 'required|email',
+            'Telefono' => 'required|string|max:100',
+            
+        
+        
+        ];
+        $mensaje=[
+        'required'=>'El :attribute es requerido',
+        ];
+        if ($request->hasFile('Foto')) {
+            $campos=['Foto' => 'required|max:10000|mimes:jpeg,png,jpg',];
+            $mensaje=['Foto.required' =>'La foto es requerida'];
+
+        }
+        $this->validate($request, $campos, $mensaje);
         //recepcionamos todos los datos menos _token _method
         $datosEmpleado = request()->except(['_token','_method']);
         // esta parte es la de la imagen
@@ -97,7 +139,8 @@ return view('empleado.edit', compact('empleado'));
 //buelve a buscar la informacion de acuerdo al id 
          $empleado=Empleado::findOrfail($id);
          //retorna al mismo formulario pero ya actualizado
-return view('empleado.edit', compact('empleado'));
+//return view('empleado.edit', compact('empleado'));
+return redirect('empleado')->with('mensaje','Empleado Modificado');
 
 
     }
@@ -111,7 +154,12 @@ return view('empleado.edit', compact('empleado'));
     public function destroy( $id)
     {
         //
-        Empleado::destroy($id);
-        return redirect('empleado');
+        $empleado=Empleado::findOrfail($id);
+        //codigo si se tiene imagenes en el registro foto es el nombre de la base de datos
+        if (Storage::delete('public/'.$empleado->foto)) {
+            Empleado::destroy($id);
+        }
+       
+        return redirect('empleado')->with('mensaje','Empleado Eliminado');
     }
 }
